# backendFilmsters/swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Filmster
  description: |
    Swagger for Filmster Api
    
    [Link to localhost:8000](http://localhost:8000/swagger.yaml)
    
    
    
  license:
    name: MIT
servers:
  - url: http://localhost:8000
paths:
  /user:
    get:
      summary: List all Users or search by query parameter
      operationId: listUsers
      tags:
        - User
      parameters:
        - in: query
          description: search for user by query parameter email
          name: email
          schema:
            type: string
      responses:
        '200':
          description: Array of Users
          content:
            application/json:    
              schema:
                $ref: '#/components/schemas/ResponseUsers'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a User
      operationId: createUser
      tags:
      - User
      requestBody:
        description: the User to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUser'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      properties:
        firstname: 
          type: string
        lastname: 
          type: string
        password: 
          type: string
        movies:
          type: array
          items:
           type: string
    ResponseUser:
      allOf:
      - $ref: '#/components/schemas/User'
      - properties:
          _id:
            type: string
            format: uuid
          __v:
            type: integer
            format: uint32  
    ResponseUsers:
      type: array
      items:
        $ref: "#/components/schemas/ResponseUser"
    Error:
      required:
        - error
      properties:
        error:
          type: object
          properties:
            expose:
              type: boolean
            statusCode:
              type: integer
              format: uint32
            status:
              type: integer
              format: uint32
            body:
              title: body as escaped json
              type: string
            type:
              title: fault desciption
              type: string